+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Fixed dirty comparison of Model attributes using the Geometry casts
+
 ## [2.0.1](https://github.com/clickbar/laravel-magellan/tree/2.0.1) - 2025-08-13
 
 ### Fixed